refactor(#1578): hashTripToken + sanitizeContext 공용 helper 추출 (중복 진짜 제거)
- src/shared/infra/monitoring/tripTokenHash.ts 신설 (Sentry SDK import 없는 순수 함수)
- device captureXEvent.ts + backend sentry.ts 양쪽 동일 helper import
- backend tsconfig include에 device tripTokenHash.ts 추가
- CPD 회피책 revert — 구조적 중복 자체 해소
